Factorized Qt4 and Qt5 modules in one base class.
